@@ -1,6 +1,6 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.1.0'
+gem 'rails', '3.2.3'
 gem 'rails-i18n'
 
 gem 'omniauth'
@@ -8,7 +8,7 @@ gem 'omniauth-twitter'
 gem 'omniauth-facebook'
 gem 'omniauth-google'
 
-gem "will_paginate", "3.0.pre2"
+gem "will_paginate", git: "https://github.com/p7r/will_paginate.git", branch: "rails3"
 gem "carrierwave"
 gem "fog", "~> 1.3.1"
 gem 'aws-s3'
@@ -18,14 +18,14 @@ gem "formtastic", "~> 2.1.1"
 gem "activeadmin", "~> 0.4.3"
 
 # Asset template engines
-gem 'sass-rails', "~> 3.1.0.rc"
+gem 'sass-rails',   '~> 3.2.3'
 gem 'coffee-script'
 gem 'uglifier'
 
 gem 'jquery-rails'
 
 group :development do
-	gem 'sqlite3', '1.3.4'
+	gem 'sqlite3'
 	gem 'faker', '0.3.1'
 end
 
