@@ -1,7 +1,7 @@
 # encoding: utf-8
 class MicropostsController < ApplicationController
-  before_filter :authenticate, :only => [:create, :destroy, :favourites, :edit]
-  before_filter :authorized_user, :only => :destroy
+  before_filter :authenticate, only: [:create, :destroy, :favourites, :edit]
+  before_filter :authorized_user, only: [:edit, :update, :destroy]
 
   def index
     @title = "microposts"
@@ -52,7 +52,7 @@ class MicropostsController < ApplicationController
     @new_micropost = Micropost.new
     if @micropost.update_attributes(params[:micropost])
       flash[:success] = "Raspa atualizada com sucesso!"
-      redirect_to @micropost.user
+      redirect_to :back
     else
       @title = "editar raspa"
       render 'edit'
@@ -61,7 +61,7 @@ class MicropostsController < ApplicationController
 
   def destroy
     @micropost.destroy
-    redirect_back_or current_user
+    redirect_to :back
   end
 
   def favouriters
