@@ -11,6 +11,7 @@
 
 class Author < ActiveRecord::Base
   attr_accessible :name, :image, :bio, :job, :tipo, :origin, :born
+  before_create :titlelize
 
   mount_uploader :image, ImageUploader
 
@@ -26,17 +27,32 @@ class Author < ActiveRecord::Base
   has_many :reverse_subscriptions, :class_name => "Subscription",
                                    :dependent => :destroy
 
-  has_many :fans, :through => :reverse_subscriptions, :source => :user
+  has_many :fans, through: :reverse_subscriptions, :source => :user
 
-  has_many :users, :through => :microposts, order: "users.created_at DESC", select: "DISTINCT users.*"
+  has_many :users, through: :microposts, order: "users.created_at DESC", select: "DISTINCT users.*"
 
   has_many :favourites
+
+  has_many :origins, through: :microposts
+
+  has_many :origins
+  has_many :books
+  has_many :poems
+  has_many :songs
+  has_many :films
+  has_many :others
 
   has_many :eleitas, through: :favourites, source: :micropost, order: "favourites.created_at DESC"
 
 	validates :name,  :presence => true,
                     :length   => { :maximum => 50 },
                     :uniqueness => { :case_sensitive => false }
+
+  def titlelize
+    if name
+      self.name = name.titlecase
+    end
+  end
 
   include PgSearch
   pg_search_scope :search, against: [:name, :job, :origin],
