@@ -11,7 +11,7 @@
 #
 
 class Micropost < ActiveRecord::Base
-  attr_accessible :content, :tag_names, :author_id, :author_name, :origem
+  attr_accessible :content, :tag_names, :author_id, :author_name, :origin_name, :origin_id, :origin_type
   
   has_many :taggings, dependent: :destroy
   has_many :tags, through: :taggings
@@ -21,6 +21,11 @@ class Micropost < ActiveRecord::Base
 
   belongs_to :user
   belongs_to :author
+
+  belongs_to :origin
+  belongs_to :book
+  belongs_to :song
+  belongs_to :other
 
   validates :content, presence: true, length: { maximum: 345 }, 
                                   uniqueness: { :case_sensitive => false }
@@ -38,9 +43,9 @@ class Micropost < ActiveRecord::Base
   scope :from_microposts_favourites_of, lambda { |micropost| favourites_of(micropost).recent }
   scope :user_feed, lambda { |user| from_users_followed_by(user).concat(from_authors_idols_of(user)).concat(from_microposts_favourites_of(user)) }
 
-  attr_accessor :author_name
-  attr_writer :tag_names, :author_name
-  before_save :assign_author, :normalize
+  attr_accessor :author_name, :origin_name, :origin_type
+  attr_writer :tag_names, :author_name, :origin_name, :origin_type
+  before_save :assign_author, :assign_origin, :normalize
   after_save :assign_tags 
 
   def normalize
@@ -90,8 +95,30 @@ class Micropost < ActiveRecord::Base
 
     def assign_author
       if author_name
-        author = Author.find_or_create_by_name(author_name)
-        self.author_id = author ? author.id : 0
+        @some_author = Author.find_or_create_by_name(author_name)
+        self.author_id = @some_author ? @some_author.id : 0
+      end
+    end
+
+    def assign_origin
+      if origin_name
+        origin = Origin.find_or_create_by_name(origin_name)
+        if origin
+          if origin.user_id.blank?
+            origin.update_attributes(user_id: user.id)
+          end
+          if !origin_type.blank?
+            origin.update_attributes(type: origin_type)
+          else
+            if origin.type.blank?
+              origin.update_attributes(type: 'Other')
+            end
+          end  
+          if !author_name.blank?
+            origin.update_attributes(author_id: @some_author.id)
+          end
+        end
+        self.origin_id = origin ? origin.id : 0
       end
     end
 
