@@ -1,5 +1,5 @@
 class RenameColumn < ActiveRecord::Migration
   def change
-    rename_column :microposts, :origin, :origin_type
+    rename_column :microposts, :origem, :origin_type
   end
 end
